third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/usecounted-features.html

<!DOCTYPE html>
<!-- This test is in chromium/ because it tests Chrome's internal
  UseCounter feature. -->
<title>Service Worker: UseCounted features</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>
<script>
function is_feature_counted(win, feature_id) {
  return win.internals.isUseCounted(win.document, feature_id);
}

function observe_use_counter(win, feature_id) {
  return win.internals.observeUseCounter(win.document, feature_id);
}

promise_test(t => {
    const kFetchEventIsReload = 2029;
    const kScript = 'resources/use-isReload-worker.js';
    const kScope = 'resources/use-isReload';
    return service_worker_unregister_and_register(t, kScript, kScope)
      .then(registration => {
          add_completion_callback(() => registration.unregister());
          const worker = registration.installing;
          return wait_for_state(t, worker, 'activated');
        })
      .then(() => {
          return with_iframe(kScope);
        })
      .then(frame => {
          return observe_use_counter(window, kFetchEventIsReload);
        })
      .then(() => {
          assert_true(is_feature_counted(window, kFetchEventIsReload));
        });
  });
</script>
